+++ libs/blocks/marquee/marquee.js
@@ -1,37 +1,10 @@
 /*
- * Copyright 2022 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-/*
  * Marquee - v6.0
  */
-import { decorateButtons, getBlockSize, applyHoverPlay } from '../../utils/decorate.js';
+
+import { applyHoverPlay, decorateButtons, getBlockSize, getVideoAttrs } from '../../utils/decorate.js';
 import { decorateBlockAnalytics, decorateLinkAnalytics } from '../../martech/attributes.js';
 import { createTag } from '../../utils/utils.js';
-
-function getAttrs(hash) {
-  const isAutoplay = hash?.includes('autoplay');
-  const isAutoplayOnce = hash?.includes('autoplay1');
-  const playOnHover = hash.includes('hoverplay');
-  if (isAutoplay && !isAutoplayOnce) {
-    return 'playsinline autoplay loop muted';
-  }
-  if (playOnHover && isAutoplayOnce) {
-    return 'playsinline autoplay muted data-hoverplay';
-  }
-  if (isAutoplayOnce) {
-    return 'playsinline autoplay muted';
-  }
-  return 'playsinline controls';
-}
 
 const decorateVideo = (container, src) => {
   if (typeof src === 'string') {
@@ -42,7 +15,7 @@ const decorateVideo = (container, src) => {
     container.classList.add('has-video');
   } else {
     const { pathname, hash } = src;
-    const attrs = getAttrs(hash);
+    const attrs = getVideoAttrs(hash);
     container.innerHTML = `<video ${attrs}>
           <source src=".${pathname}" type="video/mp4" />
         </video>`;
