+++ libs/blocks/video/video.js
@@ -1,24 +1,8 @@
-import { applyHoverPlay } from '../../utils/decorate.js';
-
-function getAttrs(hash) {
-  const isAutoplay = hash?.includes('autoplay');
-  const isAutoplayOnce = hash?.includes('autoplay1');
-  const playOnHover = hash.includes('hoverplay');
-  if (isAutoplay && !isAutoplayOnce) {
-    return 'playsinline autoplay loop muted';
-  }
-  if (playOnHover && isAutoplayOnce) {
-    return 'playsinline autoplay muted data-hoverplay';
-  }
-  if (isAutoplayOnce) {
-    return 'playsinline autoplay muted';
-  }
-  return 'playsinline controls';
-}
+import { applyHoverPlay, getVideoAttrs } from '../../utils/decorate.js';
 
 export default function init(a) {
   const { pathname, hash } = a;
-  const attrs = getAttrs(hash);
+  const attrs = getVideoAttrs(hash);
   const video = `<video ${attrs}>
         <source src=".${pathname}" type="video/mp4" />
       </video>`;
